+++ changelog.md
@@ -7,8 +7,9 @@
 
 ### Added
 - **Style Settings**: Added custom font upload support
-- **Editor Toolbar**: Added custom fonts support
 - **Style Settings**: Reset with custom confirm modal with checkboxes what to reset
+- **Editor Toolbar**: Added custom fonts support
+- **Editor Toolbar**: Added word wrap toggle button
 
 
 ## [0.6.6] - 2026-06-25